#Created new helper scripts for GUY's BRCA extractor
